Move latency stats collector to its own utility class.
